Lab 2 - Problem Solution Statement (Python for DS)
Learning Objective:
Working with data using Python libraries.
Data Visualization.
Exploratory data analysis and data preprocessing.
Building a Linear regression model to predict the tip amount based on different input features.
About the dataset (Customer Tip Data)
Dataset Source: https://www.kaggle.com/datasets/ranjeetjain3/seaborn-tips-dataset
The dataset contains information about the 244 orders served at a restaurant in the United States. Each observation includes the factors related to the order like total bill, time, the total number of people in a group, gender of the person paying for the order and so on.

Attribute Information:
total_bill: Total bill (cost of the meal), including tax, in US dollars
tip: Tip in US dollars
sex: Sex of the person paying for the meal
smoker: There is a smoker in a group or not
day: The day on which the order is served
time: Time of the order
size: Size of the group
Food servers’ tips in restaurants may be influenced by many factors, including the nature of the restaurant, the size of the party, and table locations in the restaurant. Restaurant managers need to know which factors matter when they assign tables to food servers. For the sake of staff morale, they usually want to avoid either the substance or the appearance of unfair treatment of the servers, for whom tips (at least in restaurants in the United States) are a major component of pay.